refactor(directories): move 'components' dir into 'src' dir
