+++ src/types.ts
@@ -1,0 +1,6 @@
+export interface SimplePostData {
+  title: string;
+  subhead: string;
+  slug: string;
+  date: string;
+}
